afvink42/requirements.txt
# Requirements file
# Libraries worden automatisch bij deployment geinstalleerd
# De opgeven versie is optioneel
Flask==1.0.2
Jinja2>=2.10.1
Werkzeug
mysql-connector-python
biopython==1.73